pyproject.toml
[tool.poetry]
name = "authwitchdevices"
version = "0.1.0"
description = ""
authors = ["ironfistik <ironfistik@mail.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
django = "^5.0.6"
django-user-agents = "^0.4.0"
djangorestframework = "^3.15.1"
psycopg2 = "^2.9.9"
pytest = "^8.2.1"
pytest-django = "^4.8.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
